thanks for fixing

+++ org.knime.knip.core/src/org/knime/knip/core/data/algebra/Vector.java
@@ -15,9 +15,9 @@ import net.imglib2.Point;
 
 /**
  * Class representing a simple point in the image. It mainly serves to facilitate simple operations on the coordinates.
- * 
+ *
  * @author hornm, schoenen University of Konstanz
- * 
+ *
  */
 
 public class Vector extends Point implements Cloneable {
@@ -42,11 +42,12 @@ public class Vector extends Point implements Cloneable {
      * Length (magnitude)
      */
     public long length() {
-        long l = 0;
-        // TODO overflows???
+        double l = 0;
+
         for (int i = 0; i < numDimensions(); i++) {
             l += getLongPosition(i) * getLongPosition(i);
         }
+
         return (long)Math.sqrt(l);
     }
 
@@ -60,7 +61,7 @@ public class Vector extends Point implements Cloneable {
 
     /**
      * Two norm.
-     * 
+     *
      * @return normalized vector
      */
     public Vector norm2() {
@@ -69,7 +70,7 @@ public class Vector extends Point implements Cloneable {
 
     /**
      * Addition. The points must be of same dimensionality, no check is made.
-     * 
+     *
      * @param p
      * @return the resulting point
      */
@@ -91,7 +92,7 @@ public class Vector extends Point implements Cloneable {
 
     /**
      * Multiplies a point with a scalar.
-     * 
+     *
      * @param scalar
      * @return the resulting point
      */
